% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_export.R
\name{aggStatTab}
\alias{aggStatTab}
\title{Build aggregation table}
\usage{
aggStatTab(.dat, selectSearchTerm, type = "aggregatedArticleStatistic")
}
\arguments{
\item{.dat}{The object with article data, preferably the output from function \code{\link{cleanData}}.}

\item{selectSearchTerm}{The search-term for which to create the table (defining the relevent rows of dat).}

\item{type}{The type of the aggregation table.}
}
\value{
The function returns a data frame with a aggregation of specific columns.
}
\description{
Selection of aggrageted column informations.
}
\details{
The used aggregation functions are min,median,max,mean,sd, number and frequency.
}

