% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{analysePage}
\alias{analysePage}
\title{collect metadata from webpage}
\usage{
analysePage(mediaName = "fazBlog", loadedPage)
}
\arguments{
\item{loadedResultPage}{name of saved website}
}
\value{
The function extract article attributes from query result page or article page as data frame
}
\description{
collect metadata from webpage
}

