% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_clean.R
\name{cleanCharacters}
\alias{cleanCharacters}
\title{remove false characters}
\usage{
cleanCharacters(.df)
}
\arguments{
\item{data}{frame}
}
\value{
data frame with cleaned results
}
\description{
remove false characters
}

