% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_clean.R
\name{cleanData}
\alias{cleanData}
\title{Clean article data.}
\usage{
cleanData(dat, mediaTarget)
}
\arguments{
\item{dat}{Data frame with article data, preferably the output from function \code{\link{analysePage}}.}

\item{mediaTarget}{Name of the media source (e.g. "spiegelOnline").}
}
\value{
The function returns a data frame with clean and correct formated results.
}
\description{
Cleans the article dataset.
}
\details{
This function removes unnecessary paragraphs and character from the article result set. 
It also sets the right data format (e.g. date) and encoding. In addition it parse interesting data from character vectors.
}

