% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate.R
\name{descrArticles}
\alias{descrArticles}
\title{Describe Article Data}
\usage{
descrArticles(.dat)
}
\arguments{
\item{dat}{data frame with article data}
}
\value{
data frame with additional descriptive article statistics
}
\description{
Describe Article Data
}

