% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_graph.R
\name{dualPlot}
\alias{dualPlot}
\title{Dual plot with linechart and barplot}
\usage{
dualPlot(titleMain, titleSub = "", yAxisTitle = "", xAxisTitle = "",
  refText = "", dat, pdfName, y2AxisTitle = "", selectSearchTerm)
}
\arguments{
\item{titleMain}{Main title of the plot}

\item{titleSub}{Subtitel of the plot}

\item{yAxisTitle}{Title of the y-axis}

\item{xAxisTitle}{Title of the x-axis}

\item{refText}{Name of the source of the plot}

\item{dat}{dataframe with article data}

\item{pdfName}{Name of the plot-pdf to create}

\item{y2AxisTitle}{Title of the second y-axis}

\item{selectSearchTerm}{Search-Term for which to create the plot}
}
\value{
The function query the search engine of the website and return all result-pages as htm on harddisk in the project-subfolder

pdf-file with dual plot
}
\description{
Dual plot with linechart and barplot
}

