% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_clean.R
\name{filterResultSet}
\alias{filterResultSet}
\title{remove false entries from results}
\usage{
filterResultSet(.df)
}
\arguments{
\item{data}{frame}
}
\value{
data frame with removed invalid rows
}
\description{
remove false entries from results
}

