% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate.R
\name{freqByTime}
\alias{freqByTime}
\title{aggregation: frequency by time}
\usage{
freqByTime(dat, timeDim = "\%Y", completeDim = TRUE)
}
\arguments{
\item{dat}{data frame with article data}

\item{timeDim}{dimension of time in POSIX standard format}
}
\value{
The frequency of elemtens from the given time dimension (e.g. frequency by month) as data frame

completeDim should the time dimension be continuous (TRUE/FALSE)
}
\description{
aggregation: frequency by time
}
\examples{
freqByTime(dat,"\%Y",TRUE): frequency by year
freqByTime(dat,"\%Y\%m",TRUE): frequency by year and month
}
\seealso{
See also \code{\link[base]{strptime}} for possible date conversions and formats.
}

