% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getPages}
\alias{getPages}
\title{get html}
\usage{
getPages(url, mediaName, searchTerm, pageCount)
}
\arguments{
\item{url}{URL of the Page}

\item{searchTerm}{searchTerm of url}
}
\value{
The function download the webpage for a given url.
}
\description{
get html
}

