% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getResultPagesHelper}
\alias{getResultPagesHelper}
\title{helper function for getResultPages}
\usage{
getResultPagesHelper(mediaName, searchTerm, pageCount)
}
\arguments{
\item{searchTerm}{Suuchbegriff}

\item{pageCount}{Number of Resultpage}
}
\value{
All readed html-pages are saved in a subfolder of the project
}
\description{
helper function for getResultPages
}

