% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{render}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rmarkdown}{\code{\link[rmarkdown]{render}}}
}}
\keyword{internal}

