% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper.R
\name{sampeArticles}
\alias{sampeArticles}
\title{sample articles}
\usage{
sampeArticles(dat, minArt = 10, maxArt = 30, percent = 0.05)
}
\arguments{
\item{dat}{data frame with article data}
}
\value{
data frame with sampled articles (minimum 10 articles, maximum 30 articles, otherwise 5 percent)
}
\description{
sample articles
}

