% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper.R
\name{statTab}
\alias{statTab}
\title{build export table}
\usage{
statTab(.dat, type = "articleStatistic")
}
\arguments{
\item{dat}{data frame with article data}
}
\value{
data frame with selected columns
}
\description{
build export table
}

