% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper.R
\name{aggStatTab}
\alias{aggStatTab}
\title{build aggregated export table}
\usage{
aggStatTab(.dat, selectSearchTerm, type = "aggregatedArticleStatistic")
}
\arguments{
\item{type}{export table template}

\item{dat}{data frame with article data}
}
\value{
data frame with selected columns
}
\description{
build aggregated export table
}

