% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{analysePage}
\alias{analysePage}
\title{Collect metadata of webpage.}
\usage{
analysePage(mediaName = "fazBlog", loadedPage)
}
\arguments{
\item{loadedResultPage}{Name of the website-file.}
}
\value{
The function extract article attributes from query result page or article page as data frame.
}
\description{
Analyse the Webpage and collect article attributes.
}

