% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_clean.R
\name{cleanCharacters}
\alias{cleanCharacters}
\title{Remove false characters.}
\usage{
cleanCharacters(.df)
}
\arguments{
\item{.df}{The data frame with results from DBpedia, preferably the output from function \code{\link{getConceptData}}.}
}
\value{
Data frame without metadata characters.
}
\description{
Removes special metadata characters from DBpedia-Results.
}
\examples{
\dontrun{
cleanCharacters(dat)
}
}

