% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_clean.R
\name{cleanData}
\alias{cleanData}
\title{clean article data}
\usage{
cleanData(dat, mediaTarget)
}
\arguments{
\item{mediaTarget}{name of the source media}

\item{data}{frame with article data}
}
\value{
data frame with cleane and converted results
}
\description{
clean article data
}

