% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate.R
\name{descrArticles}
\alias{descrArticles}
\title{Describe article data.}
\usage{
descrArticles(.dat)
}
\arguments{
\item{.dat}{The object with article data, preferably the output from function \code{\link{cleanData}}.}
}
\value{
The funtion returns a data frame with article descriptions like "number of words".
}
\description{
Set of article descriptions.
}
\details{
The description-columns will be added to the input data frame.
}

