% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate.R
\name{freqByTime}
\alias{freqByTime}
\title{Frequency by time}
\usage{
freqByTime(dat, timeDim = "\%Y", completeDim = TRUE)
}
\arguments{
\item{dat}{The object with article data, preferably the output from function \code{\link{cleanData}}.}

\item{timeDim}{Dimension of time in POSIX standard format.}

\item{completeDim}{If set to true, the time dimension will be coninuous (e.g. a month with no articles will appear in the frequency table).}
}
\value{
Returns the frequency of elemtens and average number of words per article for the given time dimension (e.g. frequency by month) as data frame.
}
\description{
Shows the number of articles by time (year, month, date etc.)
}
\examples{
freqByTime(dat,"\%Y",TRUE): Frequency by year
freqByTime(dat,"\%Y\%m",TRUE): Frequency by year and month
}
\seealso{
See also \code{\link[base]{strptime}} for possible date conversions and formats.
}

