% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getConceptData}
\alias{getConceptData}
\title{query DBpedia}
\usage{
getConceptData(concept = "index", subjectLabel = NULL,
  spqlEndPoint = "http://de.dbpedia.org/sparql")
}
\arguments{
\item{concept}{name of the wikipedia concept where to look after corresponding subjects (string)}

\item{subjectLabel}{if specified, additionally include matching subjects (string)}

\item{spqlEndPoint}{the sparcle endpoint where the query is send to (string)}
}
\value{
The function returns all subjects from DBpedia which associated to the given concept as data frame
}
\description{
query DBpedia
}

