% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getPages}
\alias{getPages}
\title{get html}
\usage{
getPages(url, mediaName, searchTerm, pageCount)
}
\arguments{
\item{url}{URL of the Webpage.}

\item{mediaName}{Name of digital media. Currently only Spiegel-Online, FAZ.net and Welt.de are supported.}

\item{searchTerm}{The keyword for which the search engine returned the URL.}

\item{pageCount}{Number of resultpage from search engine.}
}
\value{
The function will download the webpage for a given url as html on harddisk in the project-subfolder.
}
\description{
Get html-pages.
}
\details{
The name of the saved html-file will contain the name of the source media, the number of the resultpage the
url was found and the searchterm used for getting the url.
}

