% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getResultPages}
\alias{getResultPages}
\title{get query results}
\usage{
getResultPages(mediaName, searchTerm)
}
\arguments{
\item{searchTerm}{searchTerm}

\item{media}{name of digital media (spiegel, faz)}
}
\value{
The function query the search engine of the website and return all result-pages as htm on harddisk in the project-subfolder
}
\description{
get query results
}

