% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{getResultPagesHelper}
\alias{getResultPagesHelper}
\title{Helper function for getResultPages.}
\usage{
getResultPagesHelper(mediaName, searchTerm, pageCount)
}
\arguments{
\item{searchTerm}{The keyword the search engine should queried for.}

\item{pageCount}{Number of the Resultpage from the Query-Resultpage.}
}
\value{
The function returns all result-pages of the queried website as html on harddisk in the project-subfolder.
}
\description{
Get the results from the search engine of a digital media site for a specific searchterm beginning from page two.
}
\details{
This funtion is a helper funtion for function \code{\link{getResultPages}}. If the queried search engine
returns more than one resultpages, this function will get all pages beginning with page two.
}

