% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get.R
\name{sampeArticles}
\alias{sampeArticles}
\title{Sample articles}
\usage{
sampeArticles(dat, minArt = 10, maxArt = 30, percent = 0.05)
}
\arguments{
\item{dat}{The object with article data, preferably the output from function \code{\link{cleanData}}.}
}
\value{
The function returns a data frame with sampled articles.
}
\description{
Samples articles
}
\details{
The default sampling rules are minimum 10 articles, maximum 30 articles and otherwise 5 percent of available articles.
}
\examples{
\dontrun{
sampeArticles(dat,minArt=10,maxArt=30,percent=0.05)
sampeArticles(dat,minArt=10,maxArt=80,percent=0.20)
}
}

