% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_export.R
\name{statTab}
\alias{statTab}
\title{Arrange table.}
\usage{
statTab(.dat, type = "articleStatistic")
}
\arguments{
\item{type}{The type of the aggregation table.}

\item{dat}{The object with article data, preferably the output from function \code{\link{cleanData}}.}
}
\value{
The function returns a data frame with a specific selection of columns.
}
\description{
Selection of columns.
}

